move integration tests to the corresponding modules and add gradle conventions
