+++ buildSrc/src/main/kotlin/subproject-spring-conventions.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    id("spring-conventions")
+}
+
+tasks.withType<org.springframework.boot.gradle.tasks.bundling.BootJar> {
+    enabled = false
+}
+
+tasks.withType<org.springframework.boot.gradle.tasks.run.BootRun> {
+    enabled = false
+}
