--- domainservice/core/src/test/resources/application-test.yaml
@@ -1,40 +1,0 @@
-api:
-  finnhub:
-    token: test
-    symbols: SPUP,BNRG,HIG.PRG,RILYT,GBARF,WINN,PWWBF,RHBAF,CLVR,UPLD,AJGFF,HLLY,MSIU,ALNKF,UCRD,MGNT,IMAQ,SWRBF,BAYRY,RVHCS,APTPF,LALFS,HIVE,FNLIF,TLIS,CRTO,ORGO,OZBKF,UNIKF,DSPHF,FRCLL,CLMB,NCXS,GXRFF,JL,AMSRF,BITC,TLSNY,TGMR,TZPC,ADDHY,CBVTF,GIGM,CIBR,AFSIM,GULDF,CSBF,LFMDP,ABMC,LVWR.WS,PDEX,BFS.PRE
-
-mongock:
-  migration-scan-package: systems.ajax.malov.stockanalyzer.mongock.changelogs
-
-logExecutionTime:
-  enabled: false
-
-scheduling:
-  enabled: false
-
-nats:
-  spring:
-    io.nats.client:
-      servers: "nats://localhost:4222"
-
-spring:
-  data:
-    mongodb:
-      uri: mongodb://root:root@localhost:27017/stocks?authSource=admin
-    redis:
-      host: "localhost"
-      ttl:
-        minutes: 1
-      key:
-        prefix: stock-key-
-      timeout:
-        millis: 50
-  kafka:
-    consumer:
-      auto-offset-reset: earliest
-      properties:
-        session.timeout.ms: 10000
-        max.poll.interval.ms: 10000
-    bootstrap-servers: http://localhost:9092
-
-logging.level.org.springframework.test.context.cache: debug
