--- domainservice/src/testFixtures/kotlin/stockanalyzer/utils/StockFixture.kt
@@ -1,49 +1,0 @@
-package stockanalyzer.utils
-
-import systems.ajax.malov.stockanalyzer.stockrecord.domain.StockRecord
-import java.math.BigDecimal
-import java.time.Clock
-import java.time.Instant
-import java.time.ZoneId
-
-
-object StockFixture {
-    const val TEST_STOCK_SYMBOL = "AAPL"
-    val testDate: Instant = Clock.fixed(Instant.now(), ZoneId.of("UTC")).instant()
-
-    fun testDate(): Instant = Clock.fixed(Instant.now(), ZoneId.of("UTC")).instant()
-
-    fun domainStockRecord() = StockRecord(
-        id = null,
-        symbol = TEST_STOCK_SYMBOL,
-        openPrice = BigDecimal("1.0"),
-        highPrice = BigDecimal("1.0"),
-        lowPrice = BigDecimal("1.0"),
-        currentPrice = BigDecimal("1.0"),
-        previousClosePrice = BigDecimal("1.0"),
-        change = BigDecimal("1.0"),
-        percentChange = BigDecimal("1.0"),
-        dateOfRetrieval = testDate
-    )
-
-    fun firstPlaceStockRecord() = domainStockRecord().copy(
-        symbol = "AAPL",
-        change = BigDecimal("2.0"),
-        percentChange = BigDecimal("0.3"),
-        dateOfRetrieval = Instant.now()
-    )
-
-    fun alsoFirstPlaceStockRecord() = domainStockRecord().copy(
-        symbol = "AAPL",
-        change = BigDecimal("2.5"),
-        percentChange = BigDecimal("0.3"),
-        dateOfRetrieval = Instant.now()
-    )
-
-    fun secondPlaceStockRecord() = domainStockRecord().copy(
-        symbol = "SSSK",
-        change = BigDecimal("1.0"),
-        percentChange = BigDecimal("0.2"),
-        dateOfRetrieval = Instant.now()
-    )
-}
