+++ domainservice/stock-record/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    `spring-conventions`
+    `subproject-spring-conventions`
     `grpc-conventions`
     `testing-conventions`
 }
@@ -31,10 +31,3 @@ tasks.named("check") {
     dependsOn("detektTest")
 }
 
-tasks.withType<org.springframework.boot.gradle.tasks.bundling.BootJar> {
-    enabled = false
-}
-
-tasks.withType<org.springframework.boot.gradle.tasks.run.BootRun> {
-    enabled = false
-}
