+++ domainservice/stock-record/src/main/kotlin/systems/ajax/malov/stockanalyzer/stockrecord/infrastructure/redis/config/RedisConfig.kt
@@ -32,7 +32,7 @@ class RedisConfig(
             .timeoutOptions(
                 TimeoutOptions.builder()
                     .timeoutCommands(true)
-                    .fixedTimeout(Duration.ofMinutes(timeout))
+                    .fixedTimeout(Duration.ofMillis(timeout))
                     .build()
             ).build()
