+++ domainservice/user-tracked-symbol/src/integrationTest/resources/application-test.yaml
@@ -3,9 +3,6 @@ api:
     token: test
     symbols: SPUP,BNRG,HIG.PRG,RILYT,GBARF,WINN,PWWBF,RHBAF,CLVR,UPLD,AJGFF,HLLY,MSIU,ALNKF,UCRD,MGNT,IMAQ,SWRBF,BAYRY,RVHCS,APTPF,LALFS,HIVE,FNLIF,TLIS,CRTO,ORGO,OZBKF,UNIKF,DSPHF,FRCLL,CLMB,NCXS,GXRFF,JL,AMSRF,BITC,TLSNY,TGMR,TZPC,ADDHY,CBVTF,GIGM,CIBR,AFSIM,GULDF,CSBF,LFMDP,ABMC,LVWR.WS,PDEX,BFS.PRE
 
-mongock:
-  migration-scan-package: systems.ajax.malov.stockanalyzer.migration.mongock.changelogs
-
 logExecutionTime:
   enabled: false
 
@@ -16,14 +13,6 @@ spring:
   data:
     mongodb:
       uri: mongodb://root:root@localhost:27017/stocks?authSource=admin
-    redis:
-      host: "localhost"
-      ttl:
-        minutes: 1
-      key:
-        prefix: stock-key-
-      timeout:
-        millis: 50
   kafka:
     consumer:
       auto-offset-reset: earliest
