+++ domainservice/user-tracked-symbol/src/main/kotlin/systems/ajax/malov/stockanalyzer/usertrackedsymbol/infrastructure/kafka/producer/StockPriceNotificationProducer.kt
@@ -6,13 +6,14 @@ import reactor.kotlin.core.publisher.toFlux
 import reactor.kotlin.core.publisher.toMono
 import systems.ajax.kafka.publisher.KafkaPublisher
 import systems.ajax.malov.internalapi.KafkaTopic
+import systems.ajax.malov.stockanalyzer.usertrackedsymbol.application.port.output.StockPriceNotificationProducerOutPort
 import systems.ajax.malov.stockanalyzer.usertrackedsymbol.domain.NotificationStockPrice
 import systems.ajax.malov.stockanalyzer.usertrackedsymbol.infrastructure.kafka.mapper.NotificationStockPriceMapper.toNotificationStockPrice
 
 @Component
 class StockPriceNotificationProducer(
     private val kafkaPublisher: KafkaPublisher,
-) : systems.ajax.malov.stockanalyzer.usertrackedsymbol.application.port.output.StockPriceNotificationProducerOutPort {
+) : StockPriceNotificationProducerOutPort {
     override fun sendNotificationStockPrice(notifications: List<NotificationStockPrice>): Mono<Unit> {
         return notifications.toFlux()
             .map { it.toNotificationStockPrice() }
